+++ CHANGELOG.md
@@ -14,6 +14,13 @@
     clients and then add the new scopes `ADMIN` and `ME`. See the
     [migration notes] (docs/Migration.md#from-22-to-30) for further details.
 
+- Allow colons (:) as field separators for URNs of extensions, since this is
+  what the SCIM specification defines. Using periods (.) is still possible,
+  but will log a warning message. 
+
+- Fields of the core schemas for user and group can be fully qualified, i.e.
+  `filter=urn:ietf:params:scim:schemas:core:2.0:User:userName sw "J"` 
+
 - All invalid search queries now respond with a `400 BAD REQUEST` instead of
   `409 CONFLICT` status code.
 - Respond with `401 UNAUTHORIZED` when revoking or validating an access token